*** empty log message ***
git-svn-id: https://svn.resiprocate.org/rep/resiprocate/main/sip@347 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ resiprocate/Transport.cxx
@@ -80,7 +80,7 @@ Transport::buildFdSet( fd_set* fdSet, int* fdSetSize )
    assert( fdSetSize );
 	
    FD_SET(mFd,fdSet);
-   if ( mFd+1 > (unsigned int)(*fdSetSize) )
+   if ( mFd+1 > (*fdSetSize) )
    {
       *fdSetSize = mFd+1;
    }
